+++ src/Areas/Monitor/MonitorSetup.cs
@@ -3,6 +3,7 @@
 
 using AzureMcp.Areas.Monitor.Commands.HealthModels.Entity;
 using AzureMcp.Areas.Monitor.Commands.Log;
+using AzureMcp.Areas.Monitor.Commands.Metrics;
 using AzureMcp.Areas.Monitor.Commands.Table;
 using AzureMcp.Areas.Monitor.Commands.TableType;
 using AzureMcp.Areas.Monitor.Commands.Workspace;
@@ -19,6 +20,9 @@ public class MonitorSetup : IAreaSetup
     {
         services.AddSingleton<IMonitorService, MonitorService>();
         services.AddSingleton<IMonitorHealthModelService, MonitorHealthModelService>();
+        services.AddSingleton<IResourceResolverService, ResourceResolverService>();
+        services.AddSingleton<IMetricsQueryClientService, MetricsQueryClientService>();
+        services.AddSingleton<IMonitorMetricsService, MonitorMetricsService>();
     }
 
     public void RegisterCommands(CommandGroup rootGroup, ILoggerFactory loggerFactory)
@@ -28,11 +32,11 @@ public class MonitorSetup : IAreaSetup
         rootGroup.AddSubGroup(monitor);
 
         // Create Monitor subgroups
-        var logs = new CommandGroup("log", "Azure Monitor logs operations - Commands for querying Log Analytics workspaces using KQL.");
-        monitor.AddSubGroup(logs);
-
         var workspaces = new CommandGroup("workspace", "Log Analytics workspace operations - Commands for managing Log Analytics workspaces.");
         monitor.AddSubGroup(workspaces);
+
+        var resources = new CommandGroup("resource", "Azure Monitor resource operations - Commands for resource-centric operations.");
+        monitor.AddSubGroup(resources);
 
         var monitorTable = new CommandGroup("table", "Log Analytics workspace table operations - Commands for listing tables in Log Analytics workspaces.");
         monitor.AddSubGroup(monitorTable);
@@ -40,8 +44,17 @@ public class MonitorSetup : IAreaSetup
         var monitorTableType = new CommandGroup("type", "Log Analytics workspace table type operations - Commands for listing table types in Log Analytics workspaces.");
         monitorTable.AddSubGroup(monitorTableType);
 
+        var workspaceLogs = new CommandGroup("log", "Azure Monitor logs operations - Commands for querying Log Analytics workspaces using KQL.");
+        workspaces.AddSubGroup(workspaceLogs);
+
+        var resourceLogs = new CommandGroup("log", "Azure Monitor resource logs operations - Commands for querying resource logs using KQL.");
+        resources.AddSubGroup(resourceLogs);
+
         // Register Monitor commands
-        logs.AddCommand("query", new LogQueryCommand(loggerFactory.CreateLogger<LogQueryCommand>()));
+
+        workspaceLogs.AddCommand("query", new WorkspaceLogQueryCommand(loggerFactory.CreateLogger<WorkspaceLogQueryCommand>()));
+        resourceLogs.AddCommand("query", new ResourceLogQueryCommand(loggerFactory.CreateLogger<ResourceLogQueryCommand>()));
+
         workspaces.AddCommand("list", new WorkspaceListCommand(loggerFactory.CreateLogger<WorkspaceListCommand>()));
         monitorTable.AddCommand("list", new TableListCommand(loggerFactory.CreateLogger<TableListCommand>()));
 
@@ -54,5 +67,12 @@ public class MonitorSetup : IAreaSetup
         health.AddSubGroup(entity);
 
         entity.AddCommand("gethealth", new EntityGetHealthCommand(loggerFactory.CreateLogger<EntityGetHealthCommand>()));
+
+        // Create Metrics command group and register commands
+        var metrics = new CommandGroup("metrics", "Azure Monitor metrics operations - Commands for querying and analyzing Azure Monitor metrics.");
+        monitor.AddSubGroup(metrics);
+
+        metrics.AddCommand("query", new MetricsQueryCommand(loggerFactory.CreateLogger<MetricsQueryCommand>()));
+        metrics.AddCommand("definitions", new MetricsDefinitionsCommand(loggerFactory.CreateLogger<MetricsDefinitionsCommand>()));
     }
 }