+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 0.2.4 (Unreleased)
+## 0.2.6 (Unreleased)
 
 ### Features Added
 
@@ -9,7 +9,35 @@
 ### Bugs Fixed
 
 ### Other Changes
-Repository structure changed to organize all of an Azure service's code into a single "area" folder. [#426](https://github.com/Azure/azure-mcp/pull/426)
+
+## 0.2.5 (2025-06-26)
+
+### Bugs Fixed
+- Fixed issue where tool listing incorrectly returned resources instead of text. [#465](https://github.com/Azure/azure-mcp/issues/465)
+- Fixed invalid modification to HttpClient in KustoClient. [#433](https://github.com/Azure/azure-mcp/issues/433)
+
+## 0.2.4 (2025-06-24)
+
+### Features Added
+
+- Added new command for resource-centric logs query in Azure Monitor with command path `azmcp-monitor-resource-logs-query` - https://github.com/Azure/azure-mcp/pull/413
+- Added support for starting the server with a subset of services using the `--service` flag - https://github.com/Azure/azure-mcp/pull/424
+- Improved index schema handling in Azure AI Search (index descriptions, facetable fields, etc.) - https://github.com/Azure/azure-mcp/pull/440
+- Added new commands for querying metrics with Azure Monitor with command paths `azmcp-monitor-metrics-query` and `azmcp-monitor-metrics-definitions`. - https://github.com/Azure/azure-mcp/pull/428
+
+### Breaking Changes
+
+- Changed the command for workspace-based logs query in Azure Monitor from `azmcp-monitor-log-query` to `azmcp-monitor-workspace-logs-query`
+
+### Bugs Fixed
+
+- Fixed handling of non-retrievable fields in Azure AI Search. [#416](https://github.com/Azure/azure-mcp/issues/416)
+
+### Other Changes
+
+- Repository structure changed to organize all of an Azure service's code into a single "area" folder. ([426](https://github.com/Azure/azure-mcp/pull/426))
+- Upgraded Azure.Messaging.ServiceBus to 7.20.1 and Azure.Core to 1.46.2. ([441](https://github.com/Azure/azure-mcp/pull/441/))
+- Updated to ModelContextProtocol 0.3.0-preview1, which brings support for the 06-18-2025 MCP specification. ([431](https://github.com/Azure/azure-mcp/pull/431))
 
 ## 0.2.3 (2025-06-19)
 
@@ -141,7 +169,7 @@ Repository structure changed to organize all of an Azure service's code into a s
 ### Features Added
 
 - Support for Azure Key Vault keys https://github.com/Azure/azure-mcp/pull/119
-- Support for Kusto (Azure Data Explorer).  https://github.com/Azure/azure-mcp/pull/21
+- Support for Azure Data Explorer  https://github.com/Azure/azure-mcp/pull/21
 
 ## 0.0.13 (2025-05-06)